requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-login==0.5.0
python-dotenv==0.19.0
firebase-admin==6.4.0
bcrypt==3.2.0
python-jose==3.3.0 
functions-framework==3.0.0
gunicorn==20.1.0